move cron job to index.ts
