+++ hmns-app/server/graphql/resolvers/ButterflyResolvers.ts
@@ -53,14 +53,3 @@ Mutation: {
   },
 }
 };
-
-let randomButterflyId:string="653d56ec705f7a57e560d711"; //default BOTD
-//set new BOTD every midnight
-cron.schedule('0 0 * * *', function() {
- const randomButterfly=ButterflyResolvers.Query.randomButterfly();
- randomButterfly.then(function(result){
-    randomButterflyId=result._id.toString();
- })
- console.log(randomButterflyId);
- ButterflyResolvers.Mutation.setBOTD(null,{botdId:randomButterflyId});
-});
